pyproject.toml
[project]
name = "kaizen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgiref==3.5.2",
    "crispy-tailwind==0.5.0",
    "django==4.0.6",
    "django-ckeditor==6.7.0",
    "django-crispy-forms==1.14.0",
    "django-environ==0.9.0",
    "django-jazzmin==2.6.0",
    "djangorestframework==3.13.1",
    "mysqlclient==2.1.1",
    "pillow==9.2.0",
    "pytz==2022.1",
    "sqlparse==0.4.2",
    "tzdata==2022.1",
    "whitenoise==6.3.0",
]
